# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.ErrorResponse:
    properties:
      errorCode:
        type: string
      message:
        type: string
    type: object
...
  handler.UpdateProductRequest:
    properties:
      name:
        type: string
      unity:
        type: string
    type: object
...
  schemas.ProductResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      unity:
        type: string
      updatedAt:
        type: string
    type: object
...
host: 147.185.221.212:61489
info:
  contact:
    email: suporte@exemplo.com
    name: Equipe de Desenvolvimento
    url: https://github.com/Pmmvito/Golang-Api-Exemple
  description: "API REST para autenticação JWT, análise inteligente de notas fiscais
    de supermercado e gestão de categorias\n\n## \U0001F680 Funcionalidades Principais:\n-
    \U0001F510 **Autenticação JWT** com tokens de 7 dias\n- \U0001F916 **Análise de
    Notas com IA** usando Google Gemini (2-3s)\n- \U0001F4CA **Categorização Automática**
    de produtos de supermercado\n- \U0001F4DD **CRUD Completo** de recibos e categorias\n-
    \U0001F50D **Filtros Avançados** por categoria, data e valor\n- ⚡ **Respostas
    Otimizadas** (55% menores)\n\n## \U0001F4D6 Como Usar:\n1. Registre-se em `/register`\n2.
    Faça login em `/login` para obter o token JWT\n3. Clique em **Authorize** \U0001F513
    (cadeado verde) e cole: `Bearer SEU_TOKEN`\n4. Agora você pode testar todos os
    endpoints protegidos!\n\n## \U0001F6D2 Categorias de Supermercado (21 categorias
    distintas):\n**Básicos:** Grãos e Cereais, Massas, Padaria\n**Proteínas:** Carnes
    e Proteínas, Frios e Embutidos\n**Laticínios:** Leite, Queijos, Iogurtes, Manteiga\n**Frescos:**
    Frutas e Vegetais\n**Bebidas:** Bebidas (não alcoólicas), Bebidas Alcoólicas,
    Café e Chá\n**Congelados:** Produtos Congelados\n**Doces:** Doces e Sobremesas,
...
    Higiene Pessoal, Limpeza Doméstica, Papel e Descartáveis\n**Especiais:** Bebê
...
paths:
...
  /product/{id}:
    delete:
      consumes:
      - application/json
      - application/json
      description: |-
        Delete an existing product by its ID.
        Soft delete a product and ALL its associated receipt items across all user's receipts (sets deleted_at timestamp). Warning: This will delete ALL occurrences of this product in ALL your receipts.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found or not owned by user
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Delete product
      tags:
      - products
      - products
  /products:
    get:
      description: Lista todos os produtos cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.ProductResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Listar todos os produtos
      tags:
      - products
  /products/{id}:
    delete:
      consumes:
      - application/json
      - application/json
      description: |-
        Delete an existing product by its ID.
        Soft delete a product and ALL its associated receipt items across all user's receipts (sets deleted_at timestamp). Warning: This will delete ALL occurrences of this product in ALL your receipts.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found or not owned by user
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Delete product
      tags:
      - products
      - products
    get:
      description: Busca produto pelo ID
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ProductResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Buscar produto por ID
      tags:
      - products
    patch:
      consumes:
      - application/json
      description: Update an existing product. All fields are optional.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product data to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - products
  /products/date/{date}:
    get:
      description: Retorna todos os produtos criados em uma data específica (YYYY-MM-DD)
      parameters:
      - description: Data no formato YYYY-MM-DD
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.ProductResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Buscar produtos por data
      tags:
      - products
  /products/period:
    get:
      description: Retorna todos os produtos criados entre as query params `start`
        e `end` (RFC3339 ou YYYY-MM-DD). Ambos são obrigatórios.
      parameters:
      - description: Data/hora inicial (RFC3339 ou YYYY-MM-DD)
        in: query
        name: start
        required: true
        type: string
      - description: Data/hora final (RFC3339 ou YYYY-MM-DD)
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.ProductResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Buscar produtos por período
      tags:
      - products
...
securityDefinitions:
  BearerAuth:
    description: 'Digite "Bearer" seguido do seu token JWT (obtido no login). Exemplo:
      Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"